+++ src/DateTimeCreator.php
@@ -11,13 +11,14 @@ class DateTimeCreator
      * Returns new DateTime object formatted according to the specified format
      *
      * @param  string         $format   The format that the passed in string should be in
-     * @param  string         $time     String representing the time
+     * @param  string         $date     String representing the time
      * @return \DateTime
      * @throws Exception\InvalidDateStructureException If creation fail
      */
-    public static function createFromFormat($format, $time)
+    public static function createFromFormat($format, $date)
     {
-        if ($dateTime = \DateTime::createFromFormat($format, $time)) {
+        if ($dateTime = \DateTime::createFromFormat($format, $date)) {
+            $dateTime->setTime(0, 0, 0);
             return $dateTime;
         }